+++ posthog-react-native/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Many methods such as `capture` and `identify` no longer return the `this` object instead returning nothing
 - Fixes some typos in types
 - `shutdown` and `shutdownAsync` takes a `shutdownTimeoutMs` param with a default of 30000 (30s). This is the time to wait for flushing events before shutting down the client. If the timeout is reached, the client will be shut down regardless of pending events.
+- Adds a new `featureFlagsRequestTimeout` timeout parameter for feature flags which defaults to 3 seconds, updated from the default 10s for all other API calls.
 
 # 2.11.6 - 2024-02-22